backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
torch==2.2.2
torchvision==0.17.2
trimesh==4.0.5
pillow==10.3.0
matplotlib==3.8.4
scipy==1.12.0
requests==2.32.3

# Optional: Advanced models
# ultralytics==8.2.103
# transformers==4.44.2
# timm==1.0.9